+++ src/geocoder/mod.rs
@@ -1,12 +1,15 @@
 //! Geocoding abstraction layer.
 
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::cmp::Ordering;
+use std::sync::atomic::{AtomicU64, Ordering as AtomicOrdering};
 use std::sync::{Arc, mpsc};
 
 use calloop::channel::Event;
 use calloop::{LoopHandle, channel};
 use geocoder_nlp::SearchReference;
+use reqwest::Client;
 
+use crate::config::Config;
 use crate::geometry::GeoPoint;
 use crate::region::Regions;
 use crate::ui::view::View;
@@ -14,26 +17,34 @@ use crate::ui::view::search::SearchView;
 use crate::{Error, State};
 
 mod nlp;
+mod nominatim;
 
 /// Multi-provider geocoder.
 pub struct Geocoder {
+    nominatim_query_tx: Option<mpsc::Sender<Query>>,
     nlp_query_tx: mpsc::Sender<Query>,
+
+    result_tx: channel::Sender<(QueryId, QueryResultEvent)>,
+    nominatim_url: Arc<String>,
+    client: Client,
 
     results: Vec<QueryResult>,
     last_query: QueryId,
-    searching: bool,
+    nominatim_searching: bool,
+    nlp_searching: bool,
 }
 
 impl Geocoder {
     pub fn new(
         event_loop: LoopHandle<'static, State>,
+        config: &Config,
+        client: Client,
         regions: Arc<Regions>,
     ) -> Result<Self, Error> {
-        let (nlp_query_tx, nlp_query_rx) = mpsc::channel::<Query>();
-        let (nlp_result_tx, nlp_result_rx) = channel::channel();
+        let (result_tx, result_rx) = channel::channel();
 
         // Handle new geocoding results.
-        event_loop.insert_source(nlp_result_rx, |event, _, state| {
+        event_loop.insert_source(result_rx, |event, _, state| {
             let search_view: &mut SearchView = state.window.views.get_mut(View::Search).unwrap();
             let geocoder = search_view.geocoder_mut();
 
@@ -50,11 +61,18 @@ impl Geocoder {
                     // Add results and sort them with the best match first.
                     geocoder.results.extend(results);
                     geocoder.results.sort_unstable_by(|a, b| match (a.rank, b.rank) {
+                        (QueryResultRank::Nominatim(a), QueryResultRank::Nominatim(b)) => b.cmp(&a),
+                        (QueryResultRank::Nominatim(_), QueryResultRank::Nlp(_)) => Ordering::Less,
                         (QueryResultRank::Nlp(a), QueryResultRank::Nlp(b)) => a.total_cmp(&b),
+                        (QueryResultRank::Nlp(_), QueryResultRank::Nominatim(_)) => {
+                            Ordering::Greater
+                        },
                     });
                 },
-                // Mark current search as done.
-                QueryResultEvent::Done => geocoder.searching = false,
+                // Mark current Nominatim search as done.
+                QueryResultEvent::NominatimDone => geocoder.nominatim_searching = false,
+                // Mark current Geocoder NLP search as done.
+                QueryResultEvent::NlpDone => geocoder.nlp_searching = false,
             }
 
             search_view.update_results();
@@ -62,12 +80,32 @@ impl Geocoder {
         })?;
 
         // Spawn Geocoder NLP thread.
-        nlp::Geocoder::spawn(regions, nlp_query_rx, nlp_result_tx)?;
+        let (nlp_query_tx, nlp_query_rx) = mpsc::channel::<Query>();
+        nlp::Geocoder::spawn(regions, nlp_query_rx, result_tx.clone())?;
+
+        // Spawn Nominatim geocoder.
+        let nominatim_query_tx = if config.search.nominatim_url.is_empty() {
+            None
+        } else {
+            let (nominatim_query_tx, nominatim_query_rx) = mpsc::channel::<Query>();
+            nominatim::Geocoder::spawn(
+                client.clone(),
+                config,
+                nominatim_query_rx,
+                result_tx.clone(),
+            );
+            Some(nominatim_query_tx)
+        };
 
         Ok(Self {
+            nominatim_query_tx,
             nlp_query_tx,
+            result_tx,
+            client,
+            nominatim_url: config.search.nominatim_url.clone(),
             last_query: QueryId::new(),
-            searching: Default::default(),
+            nominatim_searching: Default::default(),
+            nlp_searching: Default::default(),
             results: Default::default(),
         })
     }
@@ -75,9 +113,13 @@ impl Geocoder {
     /// Submit a new query.
     pub fn query(&mut self, query: Query) {
         self.last_query = query.id;
-        self.searching = true;
+        self.nominatim_searching = true;
+        self.nlp_searching = true;
         self.results.clear();
 
+        if let Some(query_tx) = &self.nominatim_query_tx {
+            let _ = query_tx.send(query.clone());
+        }
         let _ = self.nlp_query_tx.send(query);
     }
 
@@ -88,18 +130,40 @@ impl Geocoder {
 
     /// Check if search is finished.
     pub fn searching(&self) -> bool {
-        self.searching
+        self.nominatim_searching || self.nlp_searching
     }
 
     /// Clear current search query and results.
     pub fn clear_results(&mut self) {
         self.last_query = QueryId::new();
-        self.searching = false;
+        self.nominatim_searching = false;
+        self.nlp_searching = false;
         self.results.clear();
+    }
+
+    /// Handle config updates.
+    pub fn update_config(&mut self, config: &Config) {
+        // Restart Nominatim geocoder on URL change.
+        if config.search.nominatim_url != self.nominatim_url {
+            self.nominatim_url = config.search.nominatim_url.clone();
+            self.nominatim_query_tx = if config.search.nominatim_url.is_empty() {
+                None
+            } else {
+                let (nominatim_query_tx, nominatim_query_rx) = mpsc::channel::<Query>();
+                nominatim::Geocoder::spawn(
+                    self.client.clone(),
+                    config,
+                    nominatim_query_rx,
+                    self.result_tx.clone(),
+                );
+                Some(nominatim_query_tx)
+            };
+        }
     }
 }
 
 /// Geocoding search query.
+#[derive(Clone)]
 pub struct Query {
     id: QueryId,
     text: String,
@@ -138,8 +202,10 @@ impl Query {
 pub enum QueryResultEvent {
     /// New query results available.
     Results(Vec<QueryResult>),
-    /// Search is done, no more results will be delivered.
-    Done,
+    /// Nominatim search is done, no more results will be delivered.
+    NominatimDone,
+    /// Geocoder NLP search is done, no more results will be delivered.
+    NlpDone,
 }
 
 /// Geocoding search result.
@@ -152,7 +218,6 @@ pub struct QueryResult {
     pub title: String,
 
     pub address: String,
-    pub postal_code: Option<String>,
 
     pub entity_type: &'static str,
 
@@ -162,7 +227,10 @@ pub struct QueryResult {
 /// Geocoder-specific search result rank.
 #[derive(Copy, Clone, Debug)]
 pub enum QueryResultRank {
+    /// Geocoder NLP result rank, lower is better.
     Nlp(f64),
+    /// Nominatim result rank, higher is better.
+    Nominatim(u32),
 }
 
 /// Unique ID of a search query.
@@ -172,6 +240,6 @@ pub struct QueryId(u64);
 impl QueryId {
     fn new() -> Self {
         static NEXT_QUERY_ID: AtomicU64 = AtomicU64::new(0);
-        Self(NEXT_QUERY_ID.fetch_add(1, Ordering::Relaxed))
+        Self(NEXT_QUERY_ID.fetch_add(1, AtomicOrdering::Relaxed))
     }
 }